moved messages under Message namespace
